Case sensitivity is now toggle-able. Strings to data points now 1 to many
Case sensitivity is now toggle-able. Default is case insensitive. A boolean
flag true in the constructor turns on case sensitivity.

In addition, strings can now be 1 to many with pieces of data. Ie, if you
supply the string ABBA twice with different data points, both pieces of
data will be captured on the node. Node.data is now an array.